+++ priam/src/main/java/com/netflix/priam/backupv2/LocalDBReaderWriter.java
@@ -20,9 +20,6 @@ package com.netflix.priam.backupv2;
 import com.google.inject.Inject;
 import com.netflix.priam.config.IConfiguration;
 import com.netflix.priam.utils.GsonJsonSerializer;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import java.io.File;
 import java.io.FileReader;
 import java.io.FileWriter;
@@ -33,15 +30,17 @@ import java.time.Instant;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
- * This class is used to create a local DB entry for SSTable components. This will be used to identify when a version
- * of a SSTable component was last uploaded or last referenced.
- * This db entry will be used to enforce the TTL of a version of a SSTable component as we will not be relying on
- * backup file system level TTL.
- * Local DB should be copied over to new instance replacing this token in case of instance replacement. If no local DB
- * is found then Priam will try to re-create the local DB using meta files uploaded to backup file system.
- * The check operation for local DB is done at every start of Priam and when operator requests to re-build the local DB.
+ * This class is used to create a local DB entry for SSTable components. This will be used to
+ * identify when a version of a SSTable component was last uploaded or last referenced. This db
+ * entry will be used to enforce the TTL of a version of a SSTable component as we will not be
+ * relying on backup file system level TTL. Local DB should be copied over to new instance replacing
+ * this token in case of instance replacement. If no local DB is found then Priam will try to
+ * re-create the local DB using meta files uploaded to backup file system. The check operation for
+ * local DB is done at every start of Priam and when operator requests to re-build the local DB.
  * Created by aagrawal on 8/31/18.
  */
 public class LocalDBReaderWriter {
@@ -54,8 +53,9 @@ public class LocalDBReaderWriter {
         this.configuration = configuration;
     }
 
-    public synchronized LocalDB upsertLocalDBEntry(final LocalDBEntry localDBEntry) throws Exception {
-        //validate the localDBEntry first
+    public synchronized LocalDB upsertLocalDBEntry(final LocalDBEntry localDBEntry)
+            throws Exception {
+        // validate the localDBEntry first
         if (localDBEntry.getTimeLastReferenced() == null)
             throw new Exception("Time last referenced in localDB can never be null");
 
@@ -68,19 +68,19 @@ public class LocalDBReaderWriter {
 
         LocalDB localDB = readAndGetLocalDB(localDBEntry.getFileUploadResult());
 
-        if (localDB == null)
-            localDB = new LocalDB(new ArrayList<>());
+        if (localDB == null) localDB = new LocalDB(new ArrayList<>());
 
-        //Verify again if someone beat you to write the entry.
+        // Verify again if someone beat you to write the entry.
         LocalDBEntry entry = getLocalDBEntry(localDBEntry.getFileUploadResult(), localDB);
-        //Write entry as it might be either -
-        //1. new component entry
-        //2. new version of file (change in compression type or file is modified e.g. stats file)
+        // Write entry as it might be either -
+        // 1. new component entry
+        // 2. new version of file (change in compression type or file is modified e.g. stats file)
         if (entry == null) {
             localDB.getLocalDBEntries().add(localDBEntry);
             writeLocalDB(localDBFile, localDB);
         } else {
-            //An entry already exists. Maybe last referenced time or backup time changed. We want to write the last time referenced.
+            // An entry already exists. Maybe last referenced time or backup time changed. We want
+            // to write the last time referenced.
             entry.setBackupTime(localDBEntry.getBackupTime());
             entry.setTimeLastReferenced(localDBEntry.getTimeLastReferenced());
             writeLocalDB(localDBFile, localDB);
@@ -106,25 +106,49 @@ public class LocalDBReaderWriter {
         return getLocalDBEntry(fileUploadResult, localDB);
     }
 
-    private LocalDBEntry getLocalDBEntry(final FileUploadResult fileUploadResult, final LocalDB localDB) throws Exception {
-        if (localDB == null || localDB.getLocalDBEntries() == null || localDB.getLocalDBEntries().isEmpty())
-            return null;
+    private LocalDBEntry getLocalDBEntry(
+            final FileUploadResult fileUploadResult, final LocalDB localDB) throws Exception {
+        if (localDB == null
+                || localDB.getLocalDBEntries() == null
+                || localDB.getLocalDBEntries().isEmpty()) return null;
 
         // Get local db entry for same file and version.
-        List<LocalDBEntry> localDBEntries = localDB.getLocalDBEntries().stream().filter(localDBEntry ->
-                // Name of the file should be same.
-                (localDBEntry.getFileUploadResult().getFileName().toFile().getName().toLowerCase().
-                        equals(fileUploadResult.getFileName().toFile().getName().toLowerCase())))
-                // Should be same version (same last modified time)
-                .filter(localDBEntry -> (localDBEntry.getFileUploadResult().getLastModifiedTime().
-                        equals(fileUploadResult.getLastModifiedTime())))
-                // Same compression as before. If we switch compression technique we can upload again.
-                .filter(localDBEntry -> (localDBEntry.getFileUploadResult().getCompression().
-                        equals(fileUploadResult.getCompression())))
-                .collect(Collectors.toList());
+        List<LocalDBEntry> localDBEntries =
+                localDB.getLocalDBEntries()
+                        .stream()
+                        .filter(
+                                localDBEntry ->
+                                        // Name of the file should be same.
+                                        (localDBEntry
+                                                .getFileUploadResult()
+                                                .getFileName()
+                                                .toFile()
+                                                .getName()
+                                                .toLowerCase()
+                                                .equals(
+                                                        fileUploadResult
+                                                                .getFileName()
+                                                                .toFile()
+                                                                .getName()
+                                                                .toLowerCase())))
+                        // Should be same version (same last modified time)
+                        .filter(
+                                localDBEntry ->
+                                        (localDBEntry
+                                                .getFileUploadResult()
+                                                .getLastModifiedTime()
+                                                .equals(fileUploadResult.getLastModifiedTime())))
+                        // Same compression as before. If we switch compression technique we can
+                        // upload again.
+                        .filter(
+                                localDBEntry ->
+                                        (localDBEntry
+                                                .getFileUploadResult()
+                                                .getCompression()
+                                                .equals(fileUploadResult.getCompression())))
+                        .collect(Collectors.toList());
 
-        if (localDBEntries.isEmpty())
-            return null;
+        if (localDBEntries.isEmpty()) return null;
 
         if (localDBEntries.size() == 1) {
             if (logger.isDebugEnabled())
@@ -133,37 +157,47 @@ public class LocalDBReaderWriter {
             return localDBEntries.get(0);
         }
 
-        throw new Exception("Unexpected behavior: More than one entry found in local database for the same file. FileUploadResult: " + fileUploadResult);
+        throw new Exception(
+                "Unexpected behavior: More than one entry found in local database for the same file. FileUploadResult: "
+                        + fileUploadResult);
     }
 
     /**
      * Gets the local db path on the local file system.
      *
-     * @param fileUploadResult This contains the SSTable component for which local db path is required.
+     * @param fileUploadResult This contains the SSTable component for which local db path is
+     *     required.
      * @return the local db path on local file system.
      */
     public Path getLocalDBPath(final FileUploadResult fileUploadResult) {
-        return Paths.get(configuration.getDataFileLocation(), LOCAL_DB,
+        return Paths.get(
+                configuration.getDataFileLocation(),
+                LOCAL_DB,
                 fileUploadResult.getKeyspaceName(),
                 fileUploadResult.getColumnFamilyName(),
-                PrefixGenerator.getSSTFileBase(fileUploadResult.getFileName().toFile().getName()) + ".localdb");
+                PrefixGenerator.getSSTFileBase(fileUploadResult.getFileName().toFile().getName())
+                        + ".localdb");
     }
 
     /**
-     * Writes the local database to the local db file. This will do a complete replace of existing local database if any.
+     * Writes the local database to the local db file. This will do a complete replace of existing
+     * local database if any.
      *
      * @param localDBFile path to the local database file.
-     * @param localDB     local database containing all the entries to the local database.
-     * @throws Exception If the path denoted is directory, write permission issues or any other exceptions.
+     * @param localDB local database containing all the entries to the local database.
+     * @throws Exception If the path denoted is directory, write permission issues or any other
+     *     exceptions.
      */
     public void writeLocalDB(final Path localDBFile, final LocalDB localDB) throws Exception {
         if (localDB == null || localDBFile == null || localDBFile.toFile().isDirectory())
-            throw new Exception("Invalid Arguments: localDbFile: " + localDBFile + ", localDB: " + localDB);
+            throw new Exception(
+                    "Invalid Arguments: localDbFile: " + localDBFile + ", localDB: " + localDB);
 
-        if (!localDBFile.getParent().toFile().exists())
-            localDBFile.getParent().toFile().mkdirs();
+        if (!localDBFile.getParent().toFile().exists()) localDBFile.getParent().toFile().mkdirs();
 
-        File tmpFile = File.createTempFile(localDBFile.toFile().getName(), ".tmp", localDBFile.getParent().toFile());
+        File tmpFile =
+                File.createTempFile(
+                        localDBFile.toFile().getName(), ".tmp", localDBFile.getParent().toFile());
         try (FileWriter writer = new FileWriter(tmpFile)) {
             writer.write(localDB.toString());
 
@@ -171,10 +205,8 @@ public class LocalDBReaderWriter {
             if (!tmpFile.renameTo(localDBFile.toFile()))
                 logger.error("Failed to persist local db: {}", localDB);
         } finally {
-            if (tmpFile != null)
-                Files.deleteIfExists(tmpFile.toPath());
+            if (tmpFile != null) Files.deleteIfExists(tmpFile.toPath());
         }
-
     }
 
     /**
@@ -182,24 +214,27 @@ public class LocalDBReaderWriter {
      *
      * @param localDBFile path the local database.
      * @return local database if file exists or empty local database.
-     * @throws Exception If there is any error in de-serializing the object or any other file system errors.
+     * @throws Exception If there is any error in de-serializing the object or any other file system
+     *     errors.
      */
     public LocalDB readLocalDB(final Path localDBFile) throws Exception {
-        //Verify file exists
-        if (!localDBFile.toFile().exists())
-            return new LocalDB(new ArrayList<>());
+        // Verify file exists
+        if (!localDBFile.toFile().exists()) return new LocalDB(new ArrayList<>());
 
         try (FileReader reader = new FileReader(localDBFile.toFile())) {
             LocalDB localDB = GsonJsonSerializer.getGson().fromJson(reader, LocalDB.class);
             String columnfamilyName = localDBFile.getParent().toFile().getName();
             String keyspaceName = localDBFile.getParent().getParent().toFile().getName();
-            localDB.getLocalDBEntries().forEach(localDBEntry -> {
-                localDBEntry.getFileUploadResult().setColumnFamilyName(columnfamilyName);
-                localDBEntry.getFileUploadResult().setKeyspaceName(keyspaceName);
-            });
+            localDB.getLocalDBEntries()
+                    .forEach(
+                            localDBEntry -> {
+                                localDBEntry
+                                        .getFileUploadResult()
+                                        .setColumnFamilyName(columnfamilyName);
+                                localDBEntry.getFileUploadResult().setKeyspaceName(keyspaceName);
+                            });
 
-            if (logger.isDebugEnabled())
-                logger.debug("Local DB: {}", localDB);
+            if (logger.isDebugEnabled()) logger.debug("Local DB: {}", localDB);
             return localDB;
         }
     }
@@ -251,7 +286,8 @@ public class LocalDBReaderWriter {
             this.backupTime = backupTime;
         }
 
-        public LocalDBEntry(FileUploadResult fileUploadResult, Instant timeLastReferenced, Instant backupTime) {
+        public LocalDBEntry(
+                FileUploadResult fileUploadResult, Instant timeLastReferenced, Instant backupTime) {
 
             this.fileUploadResult = fileUploadResult;
             this.timeLastReferenced = timeLastReferenced;
@@ -262,8 +298,5 @@ public class LocalDBReaderWriter {
         public String toString() {
             return GsonJsonSerializer.getGson().toJson(this);
         }
-
     }
-
-
 }
