+++ priam/src/test/java/com/netflix/priam/backup/TestCompression.java
@@ -17,8 +17,17 @@
 
 package com.netflix.priam.backup;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
+
 import com.netflix.priam.compress.SnappyCompression;
 import com.netflix.priam.utils.SystemUtils;
+import java.io.*;
+import java.util.Enumeration;
+import java.util.Iterator;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipFile;
+import java.util.zip.ZipOutputStream;
 import org.apache.cassandra.io.util.RandomAccessReader;
 import org.apache.commons.io.IOUtils;
 import org.junit.After;
@@ -28,26 +37,20 @@ import org.junit.Test;
 import org.xerial.snappy.SnappyInputStream;
 import org.xerial.snappy.SnappyOutputStream;
 
-import java.io.*;
-import java.util.Enumeration;
-import java.util.Iterator;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipFile;
-import java.util.zip.ZipOutputStream;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
-
 @Ignore("does this test really have to generate smoke to verify correct behavior?")
 public class TestCompression {
 
     @Before
     public void setup() throws IOException {
         File f = new File("/tmp/compress-test.txt");
-        try(FileOutputStream stream = new FileOutputStream(f)) {
+        try (FileOutputStream stream = new FileOutputStream(f)) {
             for (int i = 0; i < (1000 * 1000); i++) {
-                stream.write("This is a test... Random things happen... and you are responsible for it...\n".getBytes("UTF-8"));
-                stream.write("The quick brown fox jumps over the lazy dog.The quick brown fox jumps over the lazy dog.The quick brown fox jumps over the lazy dog.\n".getBytes("UTF-8"));
+                stream.write(
+                        "This is a test... Random things happen... and you are responsible for it...\n"
+                                .getBytes("UTF-8"));
+                stream.write(
+                        "The quick brown fox jumps over the lazy dog.The quick brown fox jumps over the lazy dog.The quick brown fox jumps over the lazy dog.\n"
+                                .getBytes("UTF-8"));
             }
         }
     }
@@ -55,8 +58,7 @@ public class TestCompression {
     @After
     public void done() {
         File f = new File("/tmp/compress-test.txt");
-        if (f.exists())
-            f.delete();
+        if (f.exists()) f.delete();
     }
 
     private void validateCompression(String uncompress, String compress) {
@@ -68,7 +70,9 @@ public class TestCompression {
     @Test
     public void zip() throws IOException {
         BufferedInputStream source = null;
-        try(ZipOutputStream out = new ZipOutputStream(new BufferedOutputStream(new FileOutputStream("/tmp/compressed.zip")))) {
+        try (ZipOutputStream out =
+                new ZipOutputStream(
+                        new BufferedOutputStream(new FileOutputStream("/tmp/compressed.zip")))) {
             byte data[] = new byte[2048];
             File file = new File("/tmp/compress-test.txt");
             FileInputStream fi = new FileInputStream(file);
@@ -89,9 +93,10 @@ public class TestCompression {
         Enumeration e = zipfile.entries();
         while (e.hasMoreElements()) {
             ZipEntry entry = (ZipEntry) e.nextElement();
-            try(BufferedInputStream is = new BufferedInputStream(zipfile.getInputStream(entry));
-                BufferedOutputStream dest1 = new BufferedOutputStream(new FileOutputStream("/tmp/compress-test-out-0.txt"), 2048)) {
-                ;
+            try (BufferedInputStream is = new BufferedInputStream(zipfile.getInputStream(entry));
+                    BufferedOutputStream dest1 =
+                            new BufferedOutputStream(
+                                    new FileOutputStream("/tmp/compress-test-out-0.txt"), 2048)) {;
                 int c;
                 byte d[] = new byte[2048];
 
@@ -108,7 +113,9 @@ public class TestCompression {
     @Test
     public void snappyCompress() throws IOException {
         FileInputStream fi = new FileInputStream("/tmp/compress-test.txt");
-        SnappyOutputStream out = new SnappyOutputStream(new BufferedOutputStream(new FileOutputStream("/tmp/test0.snp")));
+        SnappyOutputStream out =
+                new SnappyOutputStream(
+                        new BufferedOutputStream(new FileOutputStream("/tmp/test0.snp")));
         BufferedInputStream origin = new BufferedInputStream(fi, 1024);
         byte data[] = new byte[1024];
         int count;
@@ -125,7 +132,9 @@ public class TestCompression {
     @Test
     public void snappyDecompress() throws IOException {
         // decompress normally.
-        SnappyInputStream is = new SnappyInputStream(new BufferedInputStream(new FileInputStream("/tmp/test0.snp")));
+        SnappyInputStream is =
+                new SnappyInputStream(
+                        new BufferedInputStream(new FileInputStream("/tmp/test0.snp")));
         byte d[] = new byte[1024];
         FileOutputStream fos = new FileOutputStream("/tmp/compress-test-out-1.txt");
         BufferedOutputStream dest1 = new BufferedOutputStream(fos, 1024);
@@ -146,7 +155,10 @@ public class TestCompression {
         SnappyCompression compress = new SnappyCompression();
         File file = new File(new File("/tmp/compress-test.txt"), "r");
         long chunkSize = 5L * 1024 * 1024;
-        Iterator<byte[]> it = compress.compress(new AbstractBackupPath.RafInputStream(RandomAccessReader.open(file)), chunkSize);
+        Iterator<byte[]> it =
+                compress.compress(
+                        new AbstractBackupPath.RafInputStream(RandomAccessReader.open(file)),
+                        chunkSize);
         FileOutputStream ostream = new FileOutputStream("/tmp/test1.snp");
         while (it.hasNext()) {
             byte[] chunk = it.next();
@@ -159,7 +171,9 @@ public class TestCompression {
     @Test
     public void decompress() throws IOException {
         SnappyCompression compress = new SnappyCompression();
-        compress.decompressAndClose(new FileInputStream("/tmp/test1.snp"), new FileOutputStream("/tmp/compress-test-out-2.txt"));
+        compress.decompressAndClose(
+                new FileInputStream("/tmp/test1.snp"),
+                new FileOutputStream("/tmp/compress-test-out-2.txt"));
         String md1 = SystemUtils.md5(new File("/tmp/compress-test.txt"));
         String md2 = SystemUtils.md5(new File("/tmp/compress-test-out-2.txt"));
         assertEquals(md1, md2);
